test sub-amu01

+++ BIDSIFICATION/spinegeneric/fix_bids_space-other.py
@@ -26,51 +26,52 @@ def main():
         missing_files = []
         space_other_files = glob.glob(derivatives_path + '/**/*' + 'space-other' + '*.nii.gz', recursive=True)
         for label_path in space_other_files:
-            img_path = get_img_path_from_label_path(label_path)
-            if not os.path.exists(img_path):
-                img_path = get_img_path_from_label_path(label_path.replace('_space-other', ''))
-            if not os.path.exists(img_path):
-                print(f'File {img_path} is missing !')
-                missing_files.append(img_path)
-            else:
-                # Load image in RPI
-                img = Image(img_path).change_orientation('RPI')
-
-                # Load label in RPI
-                label = Image(label_path).change_orientation('RPI')
-
-                # Check if equal size
-                if img.dim[:3] != label.dim[:3]:
-                    err.append(label_path)
-                    # Move existing entity space-other after image suffix
-                    dirname = os.path.dirname(label_path)
-                    filename = os.path.basename(label_path)
-                    new_filename_split = filename.replace('_space-other', '').split('_')
-                    suffix_pos = [len(part.split('-')) for part in new_filename_split].index(1)
-                    new_filename_split.insert(suffix_pos + 1, 'space-other')
-                    new_label_path = os.path.join(dirname, "_".join(new_filename_split))
+            if 'sub-amu01' in label_path:
+                img_path = get_img_path_from_label_path(label_path)
+                if not os.path.exists(img_path):
+                    img_path = get_img_path_from_label_path(label_path.replace('_space-other', ''))
+                if not os.path.exists(img_path):
+                    print(f'File {img_path} is missing !')
+                    missing_files.append(img_path)
                 else:
-                    new_label_path = label_path.replace('_space-other', '')
+                    # Load image in RPI
+                    img = Image(img_path).change_orientation('RPI')
 
-                # Change image name
-                new_img_path = img_path.replace('_space-other','')
-                    
-                # Copy json sidecars
-                path_in_json = label_path.replace('.nii.gz', '.json')
-                path_out_json = new_label_path.replace('.nii.gz', '.json')
-                shutil.copy(path_in_json, path_out_json)
+                    # Load label in RPI
+                    label = Image(label_path).change_orientation('RPI')
 
-                # Save files using RPI orientation
-                img.save(new_img_path)
-                label.save(new_label_path)
+                    # Check if equal size
+                    if img.dim[:3] != label.dim[:3]:
+                        err.append(label_path)
+                        # Move existing entity space-other after image suffix
+                        dirname = os.path.dirname(label_path)
+                        filename = os.path.basename(label_path)
+                        new_filename_split = filename.replace('_space-other', '').split('_')
+                        suffix_pos = [len(part.split('-')) for part in new_filename_split].index(1)
+                        new_filename_split.insert(suffix_pos + 1, 'space-other')
+                        new_label_path = os.path.join(dirname, "_".join(new_filename_split))
+                    else:
+                        new_label_path = label_path.replace('_space-other', '')
 
-                # Remove files
-                print(f'Removing label NIFTII {label_path}')
-                os.remove(label_path)
-                print(f'Removing image NIFTII {img_path}')
-                os.remove(img_path)
-                print(f'Removing JSON {path_in_json}')
-                os.remove(path_in_json)
+                    # Change image name
+                    new_img_path = img_path.replace('_space-other','')
+                        
+                    # Copy json sidecars
+                    path_in_json = label_path.replace('.nii.gz', '.json')
+                    path_out_json = new_label_path.replace('.nii.gz', '.json')
+                    shutil.copy(path_in_json, path_out_json)
+
+                    # Save files using RPI orientation
+                    img.save(new_img_path)
+                    label.save(new_label_path)
+
+                    # Remove files
+                    print(f'Removing label NIFTII {label_path}')
+                    os.remove(label_path)
+                    print(f'Removing image NIFTII {img_path}')
+                    os.remove(img_path)
+                    print(f'Removing JSON {path_in_json}')
+                    os.remove(path_in_json)
                         
         print("missing files:\n" + '\n'.join(sorted(missing_files)))
         print("err files:\n" + '\n'.join(sorted(err)))
